+++ src/router/index.js
@@ -1,7 +1,7 @@
 import { createRouter, createWebHistory } from 'vue-router';
 import MainView from '@/views/main/MainView.vue';
 import MypageView from '@/views/mypage/MypageView.vue';
-import ApiDocsView from '@/views/api/ApiDocsView.vue';
+import ApiDocsView from '@/views/apidocs/ApiDocsView.vue';
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -17,7 +17,7 @@ const router = createRouter({
       component: MypageView,
     },
     {
-      path: '/api-docs',
+      path: '/docs',
       name: 'ApiDocs',
       component: ApiDocsView,
     },
